requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.1.0
autobahn==24.4.2
Automat==22.10.0
Babel==2.14.0
blinker==1.8.2
boto3==1.34.138
botocore==1.34.138
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
command-not-found==0.3
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.2.18
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==4.2.7
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-model-utils==4.5.1
django-notifications-hq==1.8.3
django-phonenumber-field==7.3.0
django-storages==1.14.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.6.4
incremental==24.7.2
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
jsonfield==3.1.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
msgpack==1.0.8
netifaces==0.11.0
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.28
Pillow==10.1.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyGObject==3.42.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==5.4.1
redis==5.0.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.2
SecretStorage==3.3.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
systemd-python==234
tomli==2.0.1
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.1.0
wadllib==1.3.6
Werkzeug==3.0.3
whitenoise==6.6.0
zipp==1.0.0
zope.interface==6.4.post2
